+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ## Unreleased
 
 ### Fixed
-
+- [#3176](https://github.com/pulumi/pulumi-kubernetes/issues/3176) ensure Helm-installed CRDs can be resolved during preview
 - [#4396](https://github.com/pulumi/pulumi-kubernetes/issues/4396) Fix panic when creating a `kubernetes.io/service-account-token` Secret with a non-existent ServiceAccount.
 - [#4394](https://github.com/pulumi/pulumi-kubernetes/issues/4394) Don't fail YAML render when a custom resource's namespace scope can't be determined offline. The resource is rendered without a namespace and a warning is logged, rather than erroring.